+++ src/lib/ogr.js
@@ -22,18 +22,21 @@ function createStream (format, options) {
 }
 
 function spawnOgr (format, options) {
+  let lastMessage
   const output = _()
   const cmd = Cmd.create(format, options)
   output.emit('log', {level: 'info', message: `Executing: OGR2OGR ${cmd.join(' ')}`})
   const ogr = spawn('ogr2ogr', cmd)
-  .on('close', (c, msg) => {
-    if (c > 0) output.emit('error', new Error(`OGR Failed: ${msg}`))
+  .on('error', e => output.emit('error', e))
+  .on('exit', (code, signal) => {
+    if (code !== 0 || signal === 'SIGKILL') output.emit('error', new Error(`OGR Failed: ${lastMessage}`))
   })
 
   _(ogr.stderr)
   .split()
   .each(data => {
     const msg = data.toString()
+    lastMessage = msg
     // Error 1: GeoJSON parsing error
     // Error 4: Failed to read GeeoJSON
     // Error 6: debug message that can be ignored
@@ -41,14 +44,13 @@ function spawnOgr (format, options) {
     if (msg.match(/ERROR\s[^6]/)) {
       ogr.stderr.unpipe()
       ogr.kill('SIGKILL')
-      ogr.emit('close', 1, msg)
     }
   })
 
   output.abort = () => {
     ogr.stderr.unpipe()
     ogr.kill('SIGKILL')
-    ogr.emit('close', 1, 'Aborted')
+    lastMessage = 'ABORTED'
   }
 
   return ogr.stdout.pipe(output)